+++ alpopca-back/src/main/java/dev/gunn96/popcat/infrastructure/web/controller/PopController.java
@@ -1,4 +1,4 @@
-package dev.gunn96.popcat.infrastructure.web;
+package dev.gunn96.popcat.infrastructure.web.controller;
 
 
 import dev.gunn96.popcat.application.dto.PopCommand;
