+++ alpopca-back/src/test/java/dev/gunn96/popcat/infrastructure/web/PopControllerTest.java
@@ -6,6 +6,7 @@ import dev.gunn96.popcat.infrastructure.geoip.GeoIpService;
 import dev.gunn96.popcat.infrastructure.security.jwt.JwtAuthenticationToken;
 import dev.gunn96.popcat.infrastructure.security.jwt.JwtProvider;
 import dev.gunn96.popcat.infrastructure.security.jwt.TokenClaims;
+import dev.gunn96.popcat.infrastructure.web.controller.PopController;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
